+++ src/amun/gpu/decoder/encoder_decoder_state.cu
@@ -8,7 +8,7 @@ namespace amunmt {
 namespace GPU {
 
 ////////////////////////////////////////////
-std::string EncoderDecoderState::Debug(size_t verbosity) const
+std::string EncoderDecoderState::Debug(unsigned verbosity) const
 {
   stringstream strm;
   strm << "states_=" << states_.Debug(verbosity);
